Cleaned Up a couple of things in AppController.
Fixed the annoying issue with closing the Preferences Window, due to the fact that the defacto Cmd-W closes SESSIONS in CoRD and not Windows (to close the preferences window with a hotkey you had to use cmd-alt-w).  Now it rearranges the two on the fly, its based on key (focus) so hopefully that is intelligent enough to not cause problems.