Changed General Preferences pane significantly, made it more Mac-ish.  We may way to play with the wording some, but it was far too verbose for a preference pane.  I moved what I'm hoping are better explanations to the help files, and provided a help button for both General Prefs and Connection Prefs.
The other large interface change here is the addition of a set of radio boxes to change modes for updating, similar to what Adium and Transmission do. I'm going off of a hybrid of Adium and T-mission's code, this will require some work done on server and maintenance side before I can really jump too far off the platform here.